Implemented the method of updating the reward by id by administrator

+++ src/main/java/com/github/discovery126/greenimpact/repository/RewardRepository.java
@@ -3,5 +3,5 @@ package com.github.discovery126.greenimpact.repository;
 import com.github.discovery126.greenimpact.model.Reward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RewardRepository extends JpaRepository<Reward, Integer> {
+public interface RewardRepository extends JpaRepository<Reward, Long> {
 }
